requirements.txt
numpy
scipy
Pillow
torch
torchvision
matplotlib
tqdm
grpcio
protobuf
grpcio-tools